move misbehaviour to submodule
